Ughh
There were 2 features cut from this update due to me being too stupid to figure them out.  When the new year starts I will be taking a break from working on the multiworld to not only focus on school but actually give it some time to breathe and let me work on documentation before going back to implementation.