ugh
Found some bugs, specifically a visual one with heroes after I patched the tower thing.  This will eventually get replaced when I randomize heroes but until then this was needed.
Also a bug related to notifications (again) but this time because I tried to access the session outside of when BloonsArchipelago was running.  Instead of just telling it to stop (probably a better solution) I just did a boolean assignment.